Added controllers for url

+++ utils/base62.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 )
 
-func Base62(n int) string {
+func Base62(n int64) string {
 	if n <= 0 {
 		return "0"
 	}
@@ -20,15 +20,15 @@ func Base62(n int) string {
 	return string(s)
 }
 
-func Base10(s string) int {
+func Base10(s string) int64 {
 	if s == "" {
 		return 0
 	}
 	var b62 = []byte("0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz")
-	var n int
-	var base int = 1
+	var n int64
+	var base int64 = 1
 	for i := len(s) - 1; i >= 0; i-- {
-		n +=base*bytes.IndexByte(b62,s[i])
+		n +=base*(int64)(bytes.IndexByte(b62,s[i]))
 		base*=62
 	}
 	return n
