+++ utils/base62_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestNumberToBase62(t *testing.T) {
 	var tests = []struct {
-		input  int
+		input  int64
 		output string
 	}{
 		{0, "0"},
@@ -25,7 +25,7 @@ func TestNumberToBase62(t *testing.T) {
 func TestBase62ToNumber(t *testing.T) {
 	var tests = []struct {
 		input  string
-		output int
+		output int64
 	}{
 		{"0", 0},
 		{"1", 1},
